requirements.txt
docx2txt==0.8
easyocr==1.4.2
gTTS==2.2.4
pandas==1.3.5
pdf2image==1.16.0
pdfplumber==0.6.2
Pillow==9.1.1
pip==22.1
pytesseract==0.3.9
python_pptx==0.6.21
streamlit==1.9.0
